revert breaking change in #234 adding env to CsmlBot Struct breaks compatibility with previous bincode encoded versions
