+++ csml_engine/src/data.rs
@@ -1,5 +1,5 @@
-use crate::{Client, Context, db_connectors, encrypt::{decrypt_data, encrypt_data}};
-use csml_interpreter::data::{CsmlBot, CsmlFlow, Message};
+use crate::{db_connectors, Client, Context};
+use csml_interpreter::data::{csml_bot::CsmlBot, Message};
 use curl::easy::Easy;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
@@ -90,173 +90,6 @@ impl BotOpt {
                 bot_version.bot
             }
         }
-    }
-}
-
-fn encrypt_env(mut env: serde_json::Value) -> Result<String, EngineError> {
-    match env.as_object_mut() {
-        Some(obj) => {
-            for (k, v) in obj.iter_mut() {
-                if !v.is_string() {
-                    return Err(EngineError::Format("env values need to be of type string".to_owned()));
-                }
-                let encrypted_value = match encrypt_data(&v) {
-                    Ok(value) => value,
-                    Err(_) => return Err(EngineError::Format(format!("Invalid [{}] value in env", k)))
-                };
-                *v = serde_json::json!(encrypted_value);
-            };
-
-            Ok(serde_json::json!(obj).to_string())
-        }
-        None => Err(EngineError::Format("Invalid env format".to_owned()))
-    }
-}
-
-fn decrypt_env(encrypted_env: String) -> Result<serde_json::Value, EngineError> {
-    let mut env = match serde_json::from_str::<serde_json::Map<String, serde_json::Value>>(&encrypted_env) {
-        Ok(value) => value,
-        Err(_e) => return Err(EngineError::Format("corrupted encrypted env value".to_owned())),
-    };
-
-    for (k, v) in env.iter_mut() {
-        if !v.is_string() {
-            return Err(EngineError::Format("env values need to be of type string".to_owned()));
-        }
-        match v.as_str() {
-            Some(value) => {
-                *v = match decrypt_data(value.to_owned()) {
-                    Ok(value) => value,
-                    Err(_) => return Err(EngineError::Format(format!("decryption fail for env value [{}]", k)))
-                };
-            }
-            None => return Err(EngineError::Format(format!("Invalid [{}] value in env", k)))
-        }
-    };
-
-    Ok(serde_json::json!(env))
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SerializeCsmlBot {
-    pub id: String,
-    pub name: String,
-    pub flows: Vec<CsmlFlow>,
-    pub native_components: Option<String>, // serde_json::Map<String, serde_json::Value>
-    pub custom_components: Option<String>, // serde_json::Value
-    pub default_flow: String,
-    pub env: Option<String>,
-}
-
-impl SerializeCsmlBot {
-    pub fn to_bot(&self) -> Result<CsmlBot, EngineError> {
-        Ok(CsmlBot {
-            id: self.id.to_owned(),
-            name: self.name.to_owned(),
-            fn_endpoint: None,
-            flows: self.flows.to_owned(),
-            native_components: {
-                match self.native_components.to_owned() {
-                    Some(value) => match serde_json::from_str(&value) {
-                        Ok(serde_json::Value::Object(map)) => Some(map),
-                        _ => unreachable!(),
-                    },
-                    None => None,
-                }
-            },
-            custom_components: {
-                match self.custom_components.to_owned() {
-                    Some(value) => match serde_json::from_str(&value) {
-                        Ok(value) => Some(value),
-                        Err(_e) => unreachable!(),
-                    },
-                    None => None,
-                }
-            },
-            default_flow: self.default_flow.to_owned(),
-            bot_ast: None,
-            env: match self.custom_components.to_owned() {
-                Some(value) => Some(decrypt_env(value)?),
-                None => None,
-            }
-        })
-    }
-
-    pub fn serialize_bot(bot: &CsmlBot) -> Result<SerializeCsmlBot, EngineError> {
-        Ok(SerializeCsmlBot {
-            id: bot.id.to_owned(),
-            name: bot.name.to_owned(),
-            flows: bot.flows.to_owned(),
-            native_components: {
-                match bot.native_components.to_owned() {
-                    Some(value) => Some(serde_json::Value::Object(value).to_string()),
-                    None => None,
-                }
-            },
-            custom_components: {
-                match bot.custom_components.to_owned() {
-                    Some(value) => Some(value.to_string()),
-                    None => None,
-                }
-            },
-            default_flow: bot.default_flow.to_owned(),
-            env: match &bot.env {
-                Some(value) => Some(encrypt_env(value.to_owned())?),
-                None => None,
-            }
-        })
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DynamoBot {
-    pub id: String,
-    pub name: String,
-    pub custom_components: Option<String>,
-    pub default_flow: String,
-    pub env: Option<String>
-}
-
-impl DynamoBot {
-    pub fn to_bot(&self, flows: Vec<CsmlFlow>) -> Result<CsmlBot, EngineError> {
-        Ok(CsmlBot {
-            id: self.id.to_owned(),
-            name: self.name.to_owned(),
-            fn_endpoint: None,
-            flows,
-            native_components: None,
-            custom_components: {
-                match self.custom_components.to_owned() {
-                    Some(value) => match serde_json::from_str(&value) {
-                        Ok(value) => Some(value),
-                        Err(_e) => unreachable!(),
-                    },
-                    None => None,
-                }
-            },
-            default_flow: self.default_flow.to_owned(),
-            bot_ast: None,
-            env: match self.env.to_owned() {
-                Some(value) => Some(decrypt_env(value.to_owned())?),
-                None => None,
-            },
-        })
-    }
-
-    pub fn serialize_bot(csml_bot: &CsmlBot) -> Result<DynamoBot, EngineError> {
-        Ok(DynamoBot {
-            id: csml_bot.id.to_owned(),
-            name: csml_bot.name.to_owned(),
-            custom_components: match csml_bot.custom_components.to_owned() {
-                Some(value) => Some(value.to_string()),
-                None => None,
-            },
-            default_flow: csml_bot.default_flow.to_owned(),
-            env: match &csml_bot.env {
-                Some(value) => Some(encrypt_env(value.to_owned())?),
-                None => None,
-            }
-        })
     }
 }
 
@@ -419,22 +252,9 @@ impl From<serde_dynamodb::Error> for EngineError {
     }
 }
 
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn ok_env_encryption() {
-        let env = serde_json::json!({
-            "key": "value",
-            "test": "toto"
-        });
-
-        let encrypted_env = encrypt_env(env.clone()).unwrap();
-
-        let decrypted_env = decrypt_env(encrypted_env).unwrap();
-
-        assert_eq!(env, decrypted_env);
-    }
-}
+// #[cfg(any(feature = "dynamo"))]
+// impl From<s3::S3Error> for EngineError {
+//     fn from(e: s3::S3Error) -> Self {
+//         EngineError::S3(e)
+//     }
+// }
