+++ csml_interpreter/examples/hello_world.rs
@@ -32,7 +32,6 @@ fn main() {
         None,
         DEFAULT_FLOW_NAME,
         None,
-        None,
     );
 
     // Create an Event
