+++ csml_interpreter/tests/generic_component.rs
@@ -42,7 +42,6 @@ fn format_message(
             .to_owned())),
         DEFAULT_FLOW_NAME,
         None,
-        None,
     );
 
     interpret(bot, context, event, None)
